feat: split loop into two sub-loops
We don't always need the code-generation node.
For tool calls where we're only running commands, for example, it is
enough for the LLM to make the tool call and pass the result to the
user. If we include the code-generation node, the LLM attempts to modify
the initial command/code even if it is not required, because the prompt
asks it to. The prompt can probably be tweaked to prevent this, but it's
cleaner to only use the code-generator node when required.